requirements.txt
fastapi[standard]
uvicorn
requests
python-dotenv